✨ Feature/homepage to postpage (#16)
* Linked cards from homepage to direct post page.

Removed date in props.

Card styles.

* Changed colorScheme and some correction in next config comments

* post: Added information about JSX transform.

* post: link text changes
