+++ .prettierignore
@@ -3,9 +3,6 @@ node_modules
 #boorstrapped components - should be removed
 pages/index_bootstrap.tsx
 
-#waitlisted files from bootstrapped repo
-lib
-
 #to be worked on files
 pages/posts
