+++ lib/api.ts
@@ -2,7 +2,7 @@ import fs from 'fs';
 import { join } from 'path';
 import matter from 'gray-matter';
 import PostType from 'types/post';
-import {OverrideProps} from 'types/utils';
+import { OverrideProps } from 'types/utils';
 
 const postsDirectory = join(process.cwd(), '_posts');
 
@@ -11,18 +11,18 @@ export function getPostSlugs() {
 }
 
 // Content is MDXRemote.source, so overriding it to string
-type RawPostType = OverrideProps<PostType, {content: string}>;
+type RawPostType = OverrideProps<PostType, { content: string }>;
 
 type PostKey = keyof PostType;
 
-export function getPostBySlug(slug: string, fields: PostKey[] = []) :  Partial<RawPostType>  {
+export function getPostBySlug(slug: string, fields: PostKey[] = []): Partial<RawPostType> {
     const realSlug = slug.replace(/\.mdx$/, '');
     const fullPath = join(postsDirectory, `${realSlug}.mdx`);
     const fileContents = fs.readFileSync(fullPath, 'utf8');
     const { data, content } = matter(fileContents);
 
     // We are just sending requested fields, hence Parital<> is used
-   const items: Partial<RawPostType> = {};
+    const items: Partial<RawPostType> = {};
 
     // Ensure only the minimal needed data is exposed
     fields.forEach((field) => {
@@ -41,11 +41,20 @@ export function getPostBySlug(slug: string, fields: PostKey[] = []) :  Partial<R
     return items;
 }
 
-export function getAllPosts(fields: PostKey[] = []) {
+export function getAllPosts(fields: PostKey[] = [], count: number = null) {
     const slugs = getPostSlugs();
-    const posts = slugs
-        .map((slug) => getPostBySlug(slug, fields))
+
+    // We need 'date' field for ordering recent posts in homepage and blogpage
+    const requestedFields: PostKey[] = fields.includes('date') ? [...fields] : [...fields, 'date'];
+
+    let posts = slugs
+        .map((slug) => getPostBySlug(slug, requestedFields))
         // sort posts by date in descending order
         .sort((post1, post2) => (post1.date > post2.date ? -1 : 1));
+
+    if (count !== null) {
+        posts.length = count;
+    }
+
     return posts;
 }
